backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.23
pydantic==2.11.7
python-multipart==0.0.20
python-dotenv==1.0.0
google-genai==1.25.0
google-adk==1.6.1
